Move table implementation to separate cpp/hpp
